feat(payment): implement subscription card skeleton and card components for payment modal
